feat: custom game creation via url props
