+++ src/routes/custom/+page.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+  import type { PageData } from "./$types";
+  import Game from "$lib/components/Game.svelte";
+  export let data: PageData;
+</script>
+
+<Game {data} />
